Refactor: Move Table and Jinja Reporter to their own module keeping backward compatibility imports
